Move proxy and WebpackConfig initialization outside of the index.js file
