+++ lib/WebpackConfig.js
@@ -20,6 +20,10 @@ const logger = require('./logger');
  */
 function validateRuntimeConfig(runtimeConfig) {
     // if you're using the encore executable, these things should never happen
+    if (null === runtimeConfig) {
+        throw new Error('RuntimeConfig must be initialized');
+    }
+
     if (null === runtimeConfig.context) {
         throw new Error('RuntimeConfig.context must be set.');
     }
@@ -29,9 +33,43 @@ function validateRuntimeConfig(runtimeConfig) {
     }
 }
 
+/**
+ * @param {RuntimeConfig} runtimeConfig
+ * @return {void}
+ */
+function checkPackageJson(runtimeConfig) {
+    // Display a warning if webpack is listed as a [dev-]dependency
+    try {
+        const packageInfo = JSON.parse(
+            fs.readFileSync(path.resolve(runtimeConfig.context, 'package.json'))
+        );
+
+        if (packageInfo) {
+            const dependencies = new Set([
+                ...(packageInfo.dependencies ? Object.keys(packageInfo.dependencies) : []),
+                ...(packageInfo.devDependencies ? Object.keys(packageInfo.devDependencies) : []),
+            ]);
+
+            if (dependencies.has('webpack')) {
+                logger.warning('Webpack is already provided by Webpack Encore, also adding it to your package.json file may cause issues.');
+            }
+        }
+    } catch (e) {
+        logger.warning('Could not read package.json file.');
+    }
+}
+
 class WebpackConfig {
-    constructor(runtimeConfig) {
+    constructor(runtimeConfig, enablePackageJsonCheck = false) {
         validateRuntimeConfig(runtimeConfig);
+
+        if (runtimeConfig.verbose) {
+            logger.verbose();
+        }
+
+        if (enablePackageJsonCheck) {
+            checkPackageJson(runtimeConfig);
+        }
 
         this.runtimeConfig = runtimeConfig;
         this.entries = new Map();
@@ -50,6 +88,8 @@ class WebpackConfig {
         this.aliases = {};
         this.externals = [];
         this.integrityAlgorithms = [];
+        this.shouldUseSingleRuntimeChunk = null;
+        this.shouldSplitEntryChunks = false;
 
         // Features/Loaders flags
         this.useVersioning = false;
@@ -96,8 +136,6 @@ class WebpackConfig {
         this.stylusLoaderOptionsCallback = () => {};
         this.babelConfigurationCallback = () => {};
         this.cssLoaderConfigurationCallback = () => {};
-        this.shouldUseSingleRuntimeChunk = null;
-        this.shouldSplitEntryChunks = false;
         this.splitChunksConfigurationCallback = () => {};
         this.watchOptionsConfigurationCallback = () => {};
         this.vueLoaderOptionsCallback = () => {};
